#END_FAIL #PROMPT  #TEST_FIX Prompt Category: Test Cases Fixes
Prompt:
The issue seems to be that it is using the same SQLite file, this is creating errors for playtests. Hence, playtests should create a new database for testing. A new issue arises get_db is not defined.

User Observation:
The model tried to redirect the DATABASE_URL to different db but syntax error arose because of whcih the test cases could not run.

Response ID:
7466069590482502674_67d9a066b09d9c26c56163fb.67d9b25ab09d9c26c561664d.67d9b25a20310add2ab61a6d